[MOICM_94] Adds a basic dashboard page
Adds the basic dashboard (with links but no search or prison selection) to allow users to navigate through the system as new pages are added. Currently only the allocation links work, the others will just reload the same page.

As part of this PR, the application.css file was renamed to application.scss to make the CSS more usable and minimal changes added for the formatting/display of the dashboard rows.

Moves the status endpoint to /status in case it is still useful.

It is expected that aas more pages are available, this page will also be changed as part of those PRs to ensure the link works.