[MOICM-111] Create base templates for POM management
Adds a template for POM management (list of POMs MOICM-112), a page for viewing a specific POM (MOICM-113) and a page for editing a single POM(MOICM-114), although the edit page does not yet support POSTing back the form to update the record.

There is an issue with the edit form, in that the designs show a cancel link next to the save button, but in this first draft the cancel link is shown as a block level element (on the line below the button).  This can be addressed when the page is completed.