Connects the links in allocations page to the right place
In allocations, the already allocated offenders are now linked up to
edit_allocate_prison_offender_managers_path, and the label for the link
temporarily changed to reallocate.  The awaiting allocation items are
linked to allocate_prison_offender_managers.

Also fixed the back links to come back to the correct tab. This may not
work in future if there is pagination involved, but we can look at
pagination then.