Adds breadcrumbs to relevant pages
Adds a new dependency (loaf - https://github.com/piotrmurach/loaf) to
make using breadcrumbs easier. Have reviewed, looks okay. You will need
to `bundle` after pulling this PR.

Breadcrumbs can be added to a particular controller as follows (taken
from allocations_controller):

```
breadcrumb 'Allocations', :allocations_path
```

If a controller adds no breadcrumbs, then none are shown.

More complex use of the breadcrumbs is possible using only: and expect:
for example in the POMS Controller we have the following where a proc
can be used for the title or the URL (which will be needed when an ID of
come sort is involved).

```
  breadcrumb 'Prison Offender Managers', :poms_path, only: [:index,
:show]
  breadcrumb -> { 'Surname, Forename' }, -> {  poms_show_path(1) },
only: [:show]
```

This will show the relevant breadcrumb items for show and index, but not
edit which is expected to have a back link instead.

Have also added a new backlink partial to make the styling easier, it
can be used in an erb in the following way:

```
<%= render :partial => "/shared/backlink", :locals => { :page =>
poms_show_path(id: 1) } %>
```