Refactors allocation controller
Rather than have a single action with a single template, this PR breaks
the page into three distinct actions (although they all currently do the
same thing) so that we can more easily control pagination and sorting.
Without this it will be hard to page as we would need to retrieve enough
offenders to know that once filtered we have enough to fill a page.