Provides a partial for pagination
Currently there is no way to navigate through pages of data that are
returned from API calls that make lists, although it is possible to
manually change the URL to replicate this.

This PR adds a partial that can be added to templates to show the
current paging state, and to navigate through the pages available (as
long as the action provides a PageMeta object). This PR doesn't
introduce any more dependencies as our needs are quite simple.

It may be that at a later stage we will need to replace the links to
?page= with a helper that replaces the parameter, for instance if there
is also likely to be a sort parameter.