Refresh VCR Cassettes
At present we have a situation that if a new test, which requires a VCR
cassette recording, to a spec file then the other cassettes need to be
refreshed to match the new time/date.  A ticket has been created for
this VCR issue to be explored as it shouldn't need us to constantly be
re-recording them.