Fetch offender details
As part of the "allocate a POM" work flow we need to display information
related to the offender needing to be allocated.  The routes and views
have been updated to display the information returned.

We have created a NullOffenderDetails object to handle the offender not
being within NOMIS and we will need to establish how to handle this on
the frontend in a following iteration.