Fixes the time requirements for VCR cassettes
Previously we had to specify the current time (using travel_to) in each
test file so that the time-sensitive HTTP requests would not have
expired.  Unfortunately, this meant we were often deleting and
recreating cassettes to handle the change in time.

This PR instead uses a before_playback hook for each VCR to set the time
to the time when the original interaction was initiated.