Adds a client for elite2 api
Currently we are using the Custody API to obtain data, but we have
agreed to move over to the Elite2 API as it is under more active
development and provides the option of more flexible endpoints.

This PR adds a client for elite2 (in the same vein as the custody
client) as well as an api class (that currently only contains a stub for
coverage purposes).  Following PRs should:

* Add offender (minimal and complete) models for elite2 data
* Add endpoints to the api class
* Fix the templates that currently use the Custody API data
* Remove the custody api and models