Get single offender info from Elite2
We are currently in the process of switching from using the Custody Api
to Elite2, as we have found it serves our purposes better.  This PR
covers the switch for getting information on a single offender, and
updates the relevant views.  This will need further work to retrieve the
release date as this no longer works in it's current form.