Fetch user details from Elite2 API
At present we are getting the details for a member of staff logging into
our service from the Custody API, but are switching over to using Elite2
and this PR covers that switch.  A subsequent PR will be raised to
remove the Custody API as will no longer be required.