Only use ECR login when pushing an image
Having it here makes its purpose clearer.

We don't need this in install_aws_cli because the aws cli is only used for
pushing docker images, and the step for that includes this line too.

We don't need this when we're deploying either - we're not doing anything
directly with ECR then, we're only telling k8s to use a different image
(which it then fetches from the ECR repo itself).