Update the allocation based routes
We are renaming 'allocate_prison_offender_managers' to 'allocate' as
this makes more sense, and is more succinct.  This involves renaming the
controller, as well as the routes.

Whilst completing this piece of work we also slightly changed the
allocate paths around the role 'new' and 'show' play, and have added a
'create' route to be able to allocate successfully.