Makes the allocation summary use a single controller action
Originally this was a single controller action, and then we changed it
to be 3 controller actions (one per tab) so that we could do paging and
breadcrumbs.

This PR changes back to a single action as we will need to obtain ALL of
the offenders in order to be able to add them to the correct tab.

As a result of this change, although pagination is rendered it won't
work particularly well as we will need to re-calculate this for each
newly calculated group.

Breadcrumbs are also a victim of this PR as they are no longer shown
without the different actions.