[WIP] Perform allocation summarisation
This PR fetches ALL of the offenders for the current prison and
categorises them according to whether they have a tier, or whether they
have been allocated or not.

It required changes to the NOMIS client in order to send a POST request
when trying to retrieve hundreds of release dates.

The general approach is to create a bucket for each group and iterate
through the offenders (which are fetched in batches), allocating them to
the correct bucket.  Each bucket has a capacity, which is the minimum
number of records to be held so that the correct page can be held.  If
we only want page 1, then we only need to hold 10 records, if we want
page 3, then we need to hold 60. Once the capacity is reached, putting
things in the bucket is a NOP.

We currently need to fetch every single offender so that we can
calculate the totals for the tabs.  Without this we could abort the
process early when all of the buckets are full. This would, however,
only give benefits when there are more than 10 in allocated, unallocated
and untiered.  Initially we would have to go through the entire list
becuase they all need a tier, and once all are tiered and allocated we'd
be in the same position as the unallocated list would be empty.

NB: We currently only get tier, we should also store case_allocation
against the OffenderShort model.

We may want to optimise the buckets to keep the last N and throw away
the previous ones as this might waste memory.