Change override_reason to be an array
It is possible for staff to choose more than one reason when overriding
the recommended POM type and so we needed to change the override table
from a string to an array.

It also made sense to change the column names in the override and
allocation table from 'override_reason' to 'override_reasons'.