Implements view list of poms and view a single POM.
The List of POMs is shown and grouped according to their status
(active/inactive).

The Single POM is shown w/o gender/email as we don't currently know
where to get these.  Functionality is added to the
PrisonOffenderManagerService which allows for the retrieval of a list of
[allocation,offender] but unfortunately in this case the offender is a
SentenceDetail as we are currently unable to get the Offender model from
/locations/description/LEI/inmates due to it not accepting a list of
IDs.