Remove the ApiResponse struct
In the Nomis::Elite2::Api we were returning the result wrapped in a
struct which had a single field.  This was because we _did_ need a paged
response that would contain the data and the page metadata.

This PR removes the ApiResponse (but leaves the ApiPaginatedResponse for
the one call that needs it) so that we no longer need to keep tacking
.data onto the end of API calls.