Request case information in batches.
Currently the case information is fetched for each offender, and so when
we are requesting all of the offenders, N at a time, we end up making N
queries to the database.

This PR requests them in batches.  So if we are processing 100 offenders
it will fetch 100 CaseInformation objects for those offenders.  This is
faster, but at the cost of more memory.  We have to do this for EVERY
offender.

I'm a little concerned about what the query plan looks like, as
initially we will be failing to find records for most of the IDs, and so
it's possible that it is slower than it should be.