Do not show inactive POMs on allocation page
When allocating a prisoner to a POM, we currently show all POMs
regardless of status. This means it is possible to allocate a prisoner
to a POM who is not available.

This PR changes the allocations_controller to only show active POMs.  To
achieve this the `get_poms` call can now take a filtering block that
will allow for get_poms to return just the POMs that caller is
interested in.  In this case, we filter with

```
get_poms('LEI') { |pom|
   pom.status == 'active'
}
```

Currently when we get_pom_detail, we find or create a PomDetail and set
the status to 'inactive' if we create a POM.  This is a stopgap, for
development purposes and will eventually be removed to not create a
PomDetail when found.  In the meantime, we've changed the default status
to 'active' rather than 'inactive'