Show validation errors when creating case information
When creating a new case information this PR will perform validation. If
validation fails then the :new action if performed and validation errors
are shown on the form.

Unfortunately because the :create call goes to /case_information/ it is
not possible to reload that form (it will instead raise a 404 because it
is looking for :index).

The messages in the error summary at the top of the page cannot link to
the relevant part of the page as a new request is made which fails with
a 404 because of the issue described above.