Implement feature to update POM
This feature allows a staff to update a POM's working pattern and
status.  This involves an edit form with radio buttons, and these
buttons are pre-selected with the POM's current details.

There is an option to make a POM inactive, and if selected this
de-allocates all the POM's cases so they can be allocated to another
POM.  A warning is displayed to avoid this option being chosen without
the person understanding this behaviour.