Implement GOVUK Notify
This PR implements basic email notification with GOVUK Notify, using an
MOJ Rails API client.  The behaviour of the notifications is to send a
new allocation email to a POM, and a re-allocation email if there was a
previous POM allocated.

Two important points:
1) At present we don't have access to POM email addresses and therefore
we have implemented a temporary 'active' flag so emails don't get sent
whilst we sort this out.

2) The emails are very basic and have had no content design, so the
templates will need to be updated once this has been done.