Add optional text field to allocation confirmation
This is an add-on to the current allocation feature and allows whoever is
doing the allocating to send an optional message to the POM.  At present
this doesn't have a word limit, but this can be added if required.  The
Notify template has been updated to include this message option and it
only displays within the email if a message is present.