Introduces a new Elite2 API call for movements
As we will need to take into account movements between prisons, and
offenders being released from prison, we will need an API call and
service to be able to do that.

The new MovementService class allows for retrieving movements on a
specific day, the intention being that you will be able to retrieve and
process these on a regular schedule performing the relevant activity
based on the Nomis::Elite2::MovementType.

MovementType.RELEASE - Deallocate offender and remove case information
MovementType.TRANSFER - Notify any currently allocated POM

The MovementService call allows for filtering the results by either the
MovementType, or MovementDirection.

Unfortunately, the API call, although returning many *description
fields, does not fill any of them in.