Allow staff without POM role to access the service
Following the implementation of the 'my caseload' and 'new allocations'
pages there was an issue that only staff with the POM role could access
all of the application.  This PR removes the reliance on people needing
to have a POM role and if they click on the aforementioned pages they
will be presented with a message confirming they don't have any
allocated cases.