Refactoring of the summary feature
As we will need to send both a page number (and eventually a sort
parameter) and only want it to apply to one tab, this PR refactors the
way that summaries are generated.  Handling different pages per tab is
messy and would be much worse with any more parameters.

In addition to separating the single action back into three, the bucket
has been simplified (and can eventually be replaced by an Array
subclass).

It is now possible in any of the summary/* urls to pass a sort
parameter, although the options are currently limited (see the Bucket
class for more).

There are no UI changes to support this feature, but it is possible to
make requests with a `sort` parameter which is "FIELD DIRECTION". For
instance all of the following should work

```
?sort=last_name asc
?sort=last_name desc
?sort=sentence_date asc
?sort=release_date asc
```

By default, all except the allocated tab now default to sorting by the
sentence_date so that the longest waiting offenders are shown first.
Specifying a sort parameter will override this. It's worth noting that
for sentence_date we sort in ASC order as we want the earliest date at
the top (we are not sorting the calculated score, just the date).