Optimise the check for 'has caseinfo'.
Currently when we are returning the list of offenders for the summary
page, we do so in blocks of N at a time.  We then send a query with a
subselect for CaseInformation objects.  This is very slow.

This PR adds a prison field to the case information, which means we can
get them all at once for a given prison (and summary page).  The
downside is that if there is an offender movement, we need to change
this field (but that is recorded in another ticket).

Once deployed to staging, we should run ..

```
update case_information set prison='LEI';
```

against the database.

Also included is a test for sorting to ensure coverage as it seems to
not have been in the last PR.