Add the support email address to the phase banner
Adds the support email address to the phase_banner file, but keeps it
secret.  Email addresses probably aren't private info, but if we can
reduce the amount of spam it has to handle, then that's a good thing.

Also adds vim swap files to the gitignore because otherwise they show up
in gitstatus if vim is still open.