Re-introduces the custody api for staff caseload lists
Currently elite2 does not provide any way of obtaining a list of
caseloads for a given user, only for the currently logged in user.

The custody API however, does provide this, so this PR provides a
custody api interface, and a single StaffApi with a single
list_caseloads method.  Given a username, it will return a list of
prison codes.  Note that this will likely include NWEB but the
associated call in the PrisonService to get the names will fail to
return a name and so it is unlikely that it will be displayed anywhere.

There is currently no service that makes use of this as we will need to
retrieve it and store it in a session somewhere, we don't want to be
making the API call on every request to the app.