Show responsibility on Make Allocations page
This is a presentation only change, so uses a helper to choose what to
display.  The use of Probation/Prison - Community/Probation is being
discussed and so there are no changes to any of the underlying
terminology.

We rely in a few places on the offender's responsibility to generate the
POM title (see Reponsibility section in allocate an offender).  This is
currently Probation POM but we don't want to change this to Community
POM based on a change of text when calculating the reponsibility.