Processes movement events from the movements API
We expect that we need to handle movements of offenders, both between
prisons and as a result of being released.

This PR is expected to be run nightly and will process the list of
movement events for the previous 24 hours.

Currently there is a hard-coded date in the rake task as we do not have
a way of creating movements on demand for testing/development purposes.
We also need some transfers to be created in legacy nomis for
dev/testing to ensure the code is correct.