Handle 204 responses from API calls
Elite2, in places, uses 204 (no content) to mean that there was
nothing to return, but that the call succeeded. This means however we do
not know whether the intended type of the response was [] or {}.  As it
would be dangerous for us to guess, instead we will return nil from
Client#get and leave it up to the caller to determine the correct type.