Stop fetching staff emails at auth
Following the implementation of fetching staff email addresses at the
auth stage, we realised that we don't actually need access to email
addresses until much later in the application.  In addition, if we get
the signed in user's email address we'll likely end up with an SPO or
OMU email address, when we really need the POM email address to send
allocation messages to.  Therefore it made sense to move this
functionality downstream. This commit covers removing the functionality
from auth, and updating the relevant specs.