Implements search
Adds a search controller and a search service that allows the user to
search for offenders by providing a string that is matched with their
first name, last name or offender number.

Currently we do this in memory because of missing functionality in
Elite2 but will often be working with offenders from the cache.  There
is no relevance as we are doing partial matches and not searching in a
corpus of documents so we can't use td/idf to determine relevance at
all.