Restrict access to the application based on the presence of a user role
To access the application, users must have the NOMIS ALLOC_MGR role set.
This PR, during the login callback, extracts the list of the roles from
the encoded token inside the encoded payload and stores them in the user
session.  The presence of the ROLE_ALLOC_MGR string is used to determine
whether the user should have access or not.

Currently the allowed? check is not done in the test environment as it
relies on encrypted data that is not available.  It _should_ be possible
to provide an encrypted token containing the roles so that tests can
complete