Improve the performance of deserialisation
On initial testing, of 100 runs, each time parsing 1000 offenders, and
then a single offender, the original deserialisation code showed:

```
    Rehearsal --------------------------------------------------
    1000 offenders  14.873416   0.102419  14.975835 ( 14.976843)
    1 offender       0.027403   0.000094   0.027497 (  0.027486)
    ---------------------------------------- total: 15.003332sec

                         user     system      total        real
    1000 offenders  14.034575   0.078000  14.112575 ( 14.112185)
    1 offender       0.018645   0.000115   0.018760 (  0.020950)
```

This, it appears, is almost entirely due to the terrible performance of
String#underscore, and making one call per field, per object.  Replacing
this call with an initial call to the plisskin gem (which will
snake_case all of the keys in a hash, or all of the keys in a list of
hashes) makes a large difference to the speed with which hashes can be
turned into models.

This PR also includes fast_underscore, which can improve the performance
of String#underscore in some cases. The end result to these changes is:

```
    Rehearsal --------------------------------------------------
    1000 offenders   6.632536   0.044988   6.677524 (  6.678995)
    1 offender       0.017296   0.000212   0.017508 (  0.017494)
    ----------------------------------------- total: 6.695032sec

                         user     system      total        real
    1000 offenders   6.557754   0.025753   6.583507 (  6.586996)
    1 offender       0.011329   0.001548   0.012877 (  0.012861)

```

This is a significant improvement, but isn't the worst culprit for
general performance as that is the multiple+sequentual API calls.  It
does however simplify some of the API classes as instead of calling .map
on a list of hashes, they can now call the `deserialise_many` method in
the ApiDeserialiser to have all the keys normalised at once.