Add more flexible way of getting all offenders
Currently we have more than one way to get lists of offenders (summary
service and search) that are complex and messy and not necessarily doing
identical things - primarily because search doesn't need all of the
information.

This PR is a bit of an experiment on how we might simplify the fetching
of lists of offenders, at the same time that we filter them and add
local data to them.

The intent is that the local API looks as follows where the arguments
passed to add_batch_filter are lambdas that receive a batch of
offenders, optionally adds data to them, and optionally removes some
records.

```
offender_list = OffenderList.new('LEI', batch_size: 200)
offender_list.add_batch_filter(add_release_date_or_remove)
offender_list.add_batch_filter(add_case_information)

offenders = offender_list.fetch
```

Longer term, and as long as the batch filters are re-entrant (
anything captured in the closure is read-only) if should in future be
possible to introduce threads in the fetch method to dramatically reduce
the cold-start time w/o the cache.