Set NOMIS active caseload
If a user signs in without an active caseload set in NOMIS there has
been an issue where they are unable to view any pages, other than the
dashboard.  This PR fixes this issue by setting the active caseload by
choosing the first prison in the caseload list, if it's not already been
set.