Filter out offenders on remand
Where an offender is on Remand (from the new convicted_status field in
OffenderShort) then exclude them from processing.  This will eventually
be the opposite, include the offender if convicted_status == 'Convicted'
but until the Elite2 change is in production, this field will be nil and
so we can't check for Convicted.

This is a much better check than just checking the absence of a release
date and we will eventually be able to filter them out on the server.