Update email fields
The Notify template has been updated with new content and therefore we
need to pass in the required fields as without it the emails aren't able
to be sent.  The addition to the template is to include whether the POM
is responsible or supporting in regards to that specific
allocated/deallocated offender.