Add additional text box to overrides page
When going through the allocation flow you have the option to override
the POM type/level.  When overriding you are directed to a page where
you are required to choose the reason(s) from a list of checkboxes.
This piece of work implements the addition of a new text box which
asks for further detail if the overriding choice is related to
suitability.