Update error messaging on overrides page
The override page allows an allocator to select one or more reasons as
to why they are overriding the recommended POM level.  This PR updates
the error messages so if a user selects an option which requires
additional information (prisoner assessed as suitable for X POM & other
detail) and they don't fill in the text box then a specific error
message is now displayed above said text box so the user knows exactly
what they need to do to pass the validations.