Make sure we get the CORRECT sentence_details
Currently when we request sentence details for an offender_no (or many)
we get back a long list of records where each offender may appear more
than once.  This makes it likely that we will be getting the wrong data.

This PR switching to useing the booking_id version of that API, so that
we can be confident that we are only getting specific sentence_details
rather than all sentence_details for previous visits.