Only show sentenced offenders
A prisoner will have had a sentence calculation and for our purposes
this means that they will either have a:
1) Release date, or
2) Parole eligibility date, or
3) HDC release date (homeDetentionCurfewEligibilityDate field).

If they do not have either of these we should be checking for a tariff
date.

Once we have all the dates we then need to display whichever is the
earliest one.

If they do not have any of these dates then we can assume they have been
convicted, but are awaiting sentencing and therefore shouldn’t be
showing up in our service.