Handle URL mangling
it's perhaps overly defensive, but currently the app crashes if somebody
changes the URL to attempt to reallocate an offender who has no active
allocation.  This PR stops them from doing that by redirecting to new
allocation instead.

It introduces active_allocation? to the AllocationService which simply
checks if there are a current active allocation for the provided
nomis_id.