Fix the case responsibility
The case responsibility is nothing to do with tier, but instead uses the
omicable status of the offender, and the length of time left in their
sentence to determine whether it is Prison/Custody or
Probation/Community.

The recommended POM should be chosen based on tier and not based on the
case ownership.  A/B is Probation, all others are Prison.