Merge the two offender models
We currently have two offender memory models, offender and
offenderShort.  This is due to the fact that there isn't one endpoint
within the elite2 api that we can gather all the relevant information
for an offender from.  However, it's not ideal having two models related
to offenders so this PR merges them into a single model.