Speed up the caseload page
The caseload page was fetching active allocations for the POM, but then
for each offender they were allocated to, was calling the get_offender
method in the OffenderService.  The only reason to call that API was to
have the case information and sentence_details set so that the pom's
responsibility could be calculated.

This PR uses a stub offender model, with the sentence_details and
case_information fetched in bulk and assigned to the fake offender
object.  This can then be used to calculate the POM's responsibility
with that offender.

The page now loads significantly faster as it reduces the number of API
calls to Elite2.