Refactor offender models
We currently have two offender memory models as we have to call two
endpoints from the API to fetch all the data we need for an offender.
There are also a number of attributes that we don't use and are storing
needlessly.  This commit removes the deserialiser's logging warning for
missed attributes and removes those attributes from both offender
models.