Expand the range of working patterns
Currently we only support 0.0, 0.2, 0.4, 0.6, 0.8 and 1.0. Our first
user went straight for 0.5.

This PR expands the options from 0.0->0.9 in 0.1 intervals, but 1.0
being handled by a separate radio button.  Selecting Full time will set
to 1.0, selecting Part time will give the option of choosing from 0 to
0.9

Unfortunately we have to jump through hoops because of the way we are
modelling POMs.  The main POM model is held in memory and we add a few
extra fields based from our local database.  The controller typically
works with the in-memory model when creating the details, but we need to
work with the subset of fields when persisting the data.  Our options
are:

* Re-write the controller to focus on working with the POMDetail model
  instead of the POM.
* Persist the POM details (name, pom_type etc) in our local database,
  and provide the ability to remove a POM from a given prison.

The latter, whilst putting more responsibility on the SPO to maintain
the list would reduce the number of API calls we would have to make and
should make a lot of the system simpler (as for instance, we can
reference allocations to a full model).