Update case owner calculation
We have been advised that whenever a POM has a supporting role then the
case will always be owned by the community, and if the POM has a
responsible role then the case will always been owned by custody.  This
PR changes the calculation to just check the outcome of the current role
and set the case owner from that.