Invert the selection for recommended POMs
aka the one where Ross fixes his stupid bug.

At some point last week, in an attempt to clean up some code, the
partitioning of POMs into recommended and non-recommended was broken.
Instead of choosing POMs that had a particular string in their
description, the code was doing the opposite.

This creates more work for users as they would then be overriding too
frequently.

This PR breaks out the logic into yet another service so that it is
testable in isolation and not just as part of the controller.  Also,
tests.