Amend Recommendation Service to use POM position
Cardiff have been setting up POMs in readiness to start using the
service, however a check in production showed an issue where the active
POM count was 12, but only 8 were displaying.  We have been using
'position_description' to filter which POMs to show, but we're not 100%
sure where this field gets set.  But, we do know that when setting up
POMs in NOMIS the LSA will have to add their 'position' and so this PR
makes a change to rely on this data as we feel we'll have more
confidence in it existing, hopefully fixing this POM list bug.