Add an ImageApi and show the offender's image
Now, when visiting /prisoners/NOMISID the offender's image is shown next
to their full name.  Unfortunately getting an offender's image in Elite2
requires an image id which is not available in the call used by
PrisonOffenderManagerService.get_offender() and so this uses the custody
API for the same effect.

A call is made to get the list of images for the offender, and this is
filtered to active images and sorted to use the most recent available
image.