Provide function to get an offender's allocation history
We need to be able to view the entire allocation history for an
offender, where the most recent (and hopefully active) allocation is top
of the list, and the historic entries are lower down the list.

This will allow us to group the allocations in windows, so that we can
show a history grouped by which Prison.  We can't use a single group_by
for that because an offender may visit a prison more than once and we
don't want historical allocations mixed up with more recent ones.