Fetch user and staff names when creating an allocation
When we create an allocation, we need to store the name of the POM and
the user who created the record as we have no guarantees it will still
exist in the API in future.  That is, if we have a username, and that
user has been deleted from NOMIS, we will not be able to look up their
name in real time.