Makes get_pom a little more reslient.
Currently in the get_pom method of PrisonOffenderManagerService there
were a couple of places where failures were not handled.  It was
possible that no poms would be retrieved, which guaranteed that the pom
filtered out of the list was nil, which guaranteed the fetching of their
email addresses failed.

This PR not fails early if there are no POMs, and then will fail early
again if the POM is not in the list of POMs for that prison.