Shows the default sort indicator
WHen visiting one of the summary pages, if no sort order is specified it
defaults to ordering by last name.  This was not showing an indicator to
show that it was sorting by last name, now it does.

At the same time the helper has been re-factored to make it simpler and
to avoid too much parsing of the current URL.