Display link to prisoner's New NOMIS profile
The new designs for post-MVP include a link on the new allocation &
reallocation pages which allows the user to view a prisoner's profile on
New Nomis.  This is a more detailed profile and makes sense that we
direct the user there, rather than duplicating the information.  Any
user that can sign into our MOIC service is able to access New Nomis so
there are no auth related issues.

The link will open a new browser tab so the user can easily switch back
to the MOIC tab and continue from where the left off.